refactor: RestaurantAddress VO record로 전환
